refactor/#6 [refactor] : Home > Card UI 변경으로 인한 로직 수정
