+++ app/src/main/java/com/devnuts/ruflu/domain/repository/HomeRepository.kt
@@ -1,7 +1,6 @@
 package com.devnuts.ruflu.domain.repository
 
-import com.devnuts.ruflu.data.api.response.card.CardModel
-import com.devnuts.ruflu.ui.model.home.UserDetailUIModel
+import com.devnuts.ruflu.data.api.response.card.UserModel
 import com.devnuts.ruflu.util.RufluApp
 import com.devnuts.ruflu.util.ServerAPI
 import retrofit2.*
@@ -10,12 +9,8 @@ class HomeRepository {
 
     private val api = RufluApp.retrofit.create(ServerAPI::class.java)
 
-    fun getUserCardList(): Call<List<CardModel>> {
+    fun getUserCardList(): Call<List<UserModel>> {
         return api.getUserCardList()
-    }
-
-    fun getNBUserList(): Call<List<UserDetailUIModel>> {
-        return api.getNBUserList()
     }
 
     fun insertHateUserCard(userCard: HashMap<String, String>): Call<String> {
@@ -24,10 +19,6 @@ class HomeRepository {
 
     fun insertLikeUserCard(userCard: HashMap<String, String>): Call<String> {
         return api.insertLikeUserCard(userCard)
-    }
-
-    fun getNBUserDtl(userId: String): Call<UserDetailUIModel> {
-        return api.getNBUserDtl(userId)
     }
 
     fun changeRatingToUser(rating: Float, userId: String): Call<String> {
