--- app/src/main/java/com/devnuts/ruflu/ui/model/home/CardImageUIModel.kt
@@ -1,9 +1,0 @@
-package com.devnuts.ruflu.ui.model.home
-
-import com.devnuts.ruflu.ui.model.CellType
-import com.devnuts.ruflu.ui.model.Model
-
-data class CardImageUIModel(
-    override val type: CellType = CellType.USER_CARD_IMAGE_CEL,
-    val image: String
-): Model(type)
